+++ src/components/CocktailDescription.css
@@ -1,32 +1,26 @@
-p{
+.TextDrink{
     text-align: center;
     margin: 0;
     font-size: 1rem;
 }
 
-h1 {
-    text-align: center;
-    font-size: 1rem;
-    margin: 0;
-    padding: 0.5rem;
-}
-
-h2 {
+.TitleDrink {
     text-align: center;
     font-size: 1rem;
     margin: 0;
     padding: 0.5rem;
 }
 
-ul{
+.ul{
     margin: 0; 
     padding: 0;
 }
 
-li {
+.li {
     list-style: none;
     text-align: center;
     margin: 0;
+    padding: 0;
     font-size: 0.8rem;
 }
