+++ src/components/CocktailPreview.jsx
@@ -1,9 +1,6 @@
 import React from 'react';
-import './Cocktails.css'
-import './makeYourCocktail.css';
-import './CocktailResult.css';
-import Axios from 'axios';
 import Cocktail from './Cocktail';
+import './CocktailResult.css';
 
 class CocktailPreview extends React.Component {
   constructor(props) {
@@ -14,14 +11,10 @@ class CocktailPreview extends React.Component {
     };
   }
 
-  
-
   clicked = () => {
     this.setState({ clicked: !this.state.clicked });
   }
-
   
-
   render() {
     return (
       <div className='DivResult'>
@@ -29,7 +22,7 @@ class CocktailPreview extends React.Component {
           <img className='ImageCocktail' src={this.props.cocktail.strDrinkThumb} alt={this.props.cocktail.strDrink} />
         </div>
         <div className="ResultDescription">
-          <h1>{this.props.cocktail.strDrink}</h1>
+          <h1 className='titlePreview'>{this.props.cocktail.strDrink}</h1>
           <button className='buttonCliked' type="button" onClick={this.clicked}>Show Cocktail description</button>
         </div>
         <div className={this.state.clicked === true ? 'modaleON' : 'modaleOFF'}>
