+++ src/components/Cocktails.css
@@ -1,3 +1,7 @@
+body{
+    background-color: #e4dabe;
+}
+
 .modaleON {
     display: inline-block;
 }
