+++ src/components/NavBar.jsx
@@ -1,21 +1,48 @@
-import React from 'react';
+import React, { useState } from "react"
+import { Link } from "react-router-dom";
+import {
+  Collapse,
+  Navbar,
+  NavbarToggler,
+  NavbarBrand,
+  Nav,
+  NavItem,
+  NavLink
+} from "reactstrap";
+import './home.css';
 import './Navbar.css';
 
-function NavBar() {
-	return (
-		<div className='DivNavBar'>
-			<nav className='NavBar'>
-				<div className='NavBarBurger'>
-					<span></span>
-					<span></span>
-					<span></span>
-				</div>
-				<div className='NavBarTitle'>
-					<h1 className="NavFontStyle">Cocktail Paradise</h1>
-				</div>
-			</nav>
-		</div>
-	);
-}
+const NavBar = props => {
+  const [isOpen, setIsOpen] = useState(false);
+  const toggle = () => setIsOpen(!isOpen);
+  return (
+    <Navbar color="light" light expand="md">
+      <NavbarBrand>
+        <Link to="/">
+          <p className='TitleNavBar'>Cocktail Paradise</p>
+        </Link>
+      </NavbarBrand>
+      <NavbarToggler onClick={toggle} />
+      <Collapse isOpen={isOpen} navbar>
+        <Nav className="mr-auto" navbar>
+          <NavItem>
+            <NavLink>
+              <Link to="/cocktails/create">
+                Make your Cocktail
+                  </Link>
+            </NavLink>
+          </NavItem>
+          <NavItem>
+            <NavLink>
+              <Link to="/cocktails">
+                Choose your Cocktail
+                  </Link>
+            </NavLink>
+          </NavItem>
+        </Nav>
+      </Collapse>
+    </Navbar>
+  );
+};
 
 export default NavBar;
