+++ src/components/Navbar.css
@@ -1,48 +1,22 @@
 @font-face {
     font-family: "Tribeca";
     src: url("/src/Fonts/Tribeca.ttf");
-   }
-
-.DivNavBar {
-    background-color: #34CFD2;
-    text-align: center;
-    margin: auto;
 }
 
-.NavFontStyle{
+.TitleNavBar {
     font-family: "Tribeca",Regular;
     margin: 0;
-    color: white;
-    font-size: 1.5rem;
+    color: rgb(37, 201, 201);
+    font-size: 1.4rem;
+    text-decoration: none;
 }
 
-.NavBar{
-    display: flex;
-    flex-direction: row;
+li{
+    text-align: center;
 }
 
-.NavBarBurger {
-    background-color: white;
-    width: 51px;
-    height: 51px;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
+li a {
+    text-decoration: none;
+    color: rgb(37, 201, 201);
+    font-size: 1.2rem;
 }
-
-span{
-    background-color: #34CFD2;
-    height: 4px;
-    width: 35px;
-    margin: 4px;
-    border-radius: 2px;
-}
-
-.NavBarTitle{
-    width: 100%;
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-}
-
