+++ src/components/Select.jsx
@@ -1,4 +1,10 @@
 import React from 'react';
+import {
+  Label,
+  Input,
+  FormGroup
+} from "reactstrap";
+import './makeYourCocktail.css';
 
 class Select extends React.Component {
 
@@ -12,25 +18,29 @@ class Select extends React.Component {
 
 
   render() {
-    return(
-      <div>
-      <div>
-        <p>Ingredient 1</p>
-        <label htmlFor="ingredient-select" />
-        <select name="ingredients1" id="ingredient1-select" onChange={(event) => this.seachInput1(event)}>
-          <option value="">--Please choose your ingredient</option>
-          {this.props.list.map((ingredient) => <option value={ingredient}>{ingredient}</option>) }
-        </select>
+    return (
+      <div className="flexSearchBar">
+        <div>
+          <p>Ingredient 1</p>
+          <FormGroup>
+            <Label htmlFor="ingredient-select" />
+            <Input type="select" className='SelectOption' name="ingredients1" id="ingredient1-select" onChange={(event) => this.seachInput1(event)}>
+              <option value="">--Please choose your ingredient</option>
+              {this.props.list.map((ingredient) => <option value={ingredient}>{ingredient}</option>)}
+            </Input>
+          </FormGroup>
+        </div>
+        <div>
+          <p>Ingredient 2</p>
+          <FormGroup>
+            <Label htmlFor="ingredient-select" />
+            <Input type="select" className='SelectOption' name="ingredients2" id="ingredient2-select" onChange={(event) => this.seachInput2(event)}>
+              <option value="">--Please choose your ingredient</option>
+              {this.props.list.map((ingredient) => <option value={ingredient}>{ingredient}</option>)}
+            </Input>
+          </FormGroup>
+        </div>
       </div>
-      <div>
-      <p>Ingredient 2</p>
-        <label htmlFor="ingredient-select" />
-        <select name="ingredients2" id="ingredient2-select" onChange={(event) => this.seachInput2(event)}>
-          <option value="">--Please choose your ingredient</option>
-          {this.props.list.map((ingredient) => <option value={ingredient}>{ingredient}</option>) }
-        </select>
-      </div>
-    </div>
     )
   }
 }
